refactor(tests): extract setupDuplicateDetectionWithExisting, compress vi.mock factories, merge alternateTopics into validation, remove unused existing vars (-111 LOC)
